+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+### Added
+- yarpctest: Add `ContextWithCall` function to ease testing of functions that
+  use `yarpc.CallFromContext`.
 
 ## [1.44.0] - 2020-02-27
 ### Added